More the karp-cli to another crate.
Muuuuch cleaner way of dealing with features.
