+++ moq-karp-cli/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "moq-karp-cli"
+description = "Media over QUIC"
+authors = ["Luke Curley <kixelated@gmail.com>"]
+repository = "https://github.com/kixelated/moq-rs"
+license = "MIT OR Apache-2.0"
+
+version = "0.1.0"
+edition = "2021"
+
+keywords = ["quic", "http3", "webtransport", "media", "live"]
+categories = ["multimedia", "network-programming", "web-programming"]
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+moq-karp = { path = "../moq-karp", version = "0.15" }
+
+moq-native = { path = "../moq-native", version = "0.6" }
+clap = { version = "4", features = ["derive"] }
+anyhow = { version = "1", features = ["backtrace"] }
+
+axum = { version = "0.8", features = ["tokio"] }
+hyper-serve = { version = "0.6", features = ["tls-rustls"] }
+tower-http = { version = "0.6", features = ["cors"] }
+
+tokio = { version = "1.43", features = ["full"] }
+url = "2"
+tracing = "0.1"
+
+[[bin]]
+name = "moq-karp"
+path = "src/main.rs"
