+++ moq-karp-cli/src/fingerprint.rs
